Better performance in different base page reconfiguration
Modified GetProdFamily.jsp to eliminate select list and instead display
the product family data within a table to be selected. Minor
modifications to ProcessProdFamilyAction.java, adding a hyphen within
the display method to parse the map value within the jsp and eliminated
the need to parse the selectedProdFamily value to obtain the product
number since the number is passed through a hidden tag when the row is
selected and the form is submitted.