Increase Code Coverage and Improve Testing For Label Printing
SAST flagged the original testing methods as blockers due to the lack of
assertions in the test cases. Additionally, the original testing methods
were encountering a null pointer exception when instantiating the
service class objects needed for the label printing class. This was
significantly reducing the amount of code that was being covered
initially by the tests.

A boolean class variable has been created and is returned to determine
if the creation of the label is successful and the instantiation of the
service class objects is done in each test case.