SonarQube 'Add th headers to table' issue fix
An empty table element was placed on the color lookup modal which gets
populated when a color lookup search is done. However, since it's an
empty table and it lacks 'th' elements. Sonar flagged this as a major
issue for assistive technologies.

The table element is now created when a user performs their first color
lookup. Further inspection of the elements created by the DataTables API
shows that the table contains the proper elements for these
technologies.