Parameterize Action Tests
SonarQube issue fix 'Replace these tests with a single parameterized
one'. This major issue wants the test case to use JUnit 5's
ParameterizedTest and ValueSource annotations to pass values that would
replace the hardcoded executeAction values.

To perform the same action with JUnit 4. It requires the class to
replace SpringJUnit4ClassRunner.class value with Parameterized.class in
the RunWith annotation and create SpringClassRules and SpringMethodRules
to provide the functionality of the TestContextManager and access to the
ApplicationContext.