SC Feature - Formula Display - Move Hand Dispense to lower priority
Hand dispense button moved to second row. Next job button moved closer
to the other buttons on the first row.

Dispense/Save button will now be displayed as the 'primary' if the
account has a tinter connected/not connected. The enter key will also
trigger action of the button that is the 'primary'.