prePrintSave to use different validation method for QtyOrdered
The prePrintSave function performs a save so that the control number and
line number of the order is populated and they can be displayed on the
label. The call to validateQtyOrdered calls another function called
saveQtyOrdered. This contains an Ajax call to set the qtyOrdered field
in the action class and sets a flag indicating that the record is
'dirty' and needs to be saved.

This field is already validated and saved when the onblur event of the
qtyOrdered field is triggered. We only need to validate that a
legitimate quantity is entered.