SonarQube Adjustments
Reduced duplicate code in CorrectFormula.jsp by adding a
saveCorrectionStep function that performs the ajax call to write them
for both hand dispense and automated dispenses.

Adjusted convertFormulaToDispenseItems method to return ERROR or SUCCESS
immediately over saving the value in a variable to be returned at the
end.

Added code coverage for the ProcessCorrectFormulaAction class.