Increase Logging to help capture cause for null FormulationResponse
A null FormulationResponse object was captured when a customer was
performing a reformulation of an existing custom formula. Due to the
lack of color/product/control number information that caused the issue.
Recreation has not been possible despite numerous attempts. In addition,
there are two places where this object can potentially be null which
further complicates the issue.