Fix ProcessProductLookupAction's call to ProductService's compatableBase
Recent changes to ProcutService.productAutocompleteCompatibleBase and
cdsProdDao.listForAutocompleteProductCompatibleBase can no longer accept
an empty string for the partialProductId field due to the way the query
is dynamically generated. Since this Action simply wants a full set of
products based on the bases that the color prefers. We will pass it a %
so the new dynamic query can build while also returning the full list
that we want when we land on DetProduct.jsp.