SCD-680 - New Product Validation Implementation
The validation process has been broken into two distinct parts.
Validating the products existence in the master product table and
determining if the size code the product is currently defined in
Sher-Color. Products whose size code is not defined will only be allowed
when creating custom manual colors. A warning message will be created
and sent to the user notifying them of the unavailable tinting
information of the specified product.