Set Request Object's Light Source to Null on reset
The ProcessFormulaAction.estimateRgbHex method checks for a null light
source value and will default to using D65 as the source. This will
prevent imported custom manual transactions on SCW from failing when it
goes to generate a RgbHex value to display it as a color in the swatch
circle on the display formula page.